finish FE changes from rebase and refactor details modal under new folder
